+++ CHANGELOG.md
@@ -13,6 +13,7 @@ API changes:
 - Added true volumetric collision detection for spell projectiles
 - Added server-side entry point for entity spell casting (SpellHelper.java)
 - Added some new assets: player animations, particles, sounds
+- Added new external spell school `PHYSICAL_MELEE_DUAL`, providing the combined power value of main hand and offhand melee attacks
 - Spell data structure changes:
   - Added new root field `order`, for sorting spells on the hotbar and spell binding GUI
 - Spell execution pipeline now supports performing impacts on other's behalf (ImpactContext.effectiveCaster field)